+++ ModLoader/ModEntry.cs
@@ -31,7 +31,7 @@ namespace Sheepy.Modnix {
       public bool Disabled;
       public SourceLevels? LogLevel;
       public long? LoadIndex;
-      public bool IsDefaultSettings => ! Disabled && LogLevel == null && LoadIndex == null;
+      public bool IsDefaultSettings => ! Disabled && ! LogLevel.HasValue && ! LoadIndex.HasValue;
    }
 
    public class ModEntry : ModSettings {