+++ ModLoader/ModScanner.cs
@@ -122,10 +122,7 @@ namespace Sheepy.Modnix {
             return null;
          }
          Log.Info( "Found mod {0} at {1} ({2} dlls)", meta.Id, file, meta.Dlls?.Length ?? 0 );
-         var result = new ModEntry( file, meta );
-         if ( result.LogLevel < ModLoader.Settings.LogLevel )
-            result.LogLevel = ModLoader.Settings.LogLevel;
-         return result;
+         return new ModEntry( file, meta );
       } catch ( Exception ex ) { Log.Warn( ex ); return null; } }
 
       private static ModMeta ParseInfoJs ( string js, string default_id = null ) { try {
@@ -314,21 +311,28 @@ namespace Sheepy.Modnix {
          EnabledMods.Clear();
          EnabledMods.AddRange( AllMods.Where( e => ! e.Disabled ) );
          Log.Info( "Resolving {0} mods", EnabledMods.Count );
+         ApplyUserOverride();
          EnabledMods.Sort( CompareModIndex );
-         RemoveDisabledMods();
          RemoveDuplicateMods();
          RemoveUnfulfilledMods();
          RemoveConflictMods();
       }
 
-      private static void RemoveDisabledMods () {
+      private static void ApplyUserOverride () {
          var settings = ModLoader.Settings.Mods;
          if ( settings == null ) return;
-         Log.Verbo( "Check manually disabled mods" );
+         Log.Verbo( "Check manual mod settings" );
          foreach ( var mod in EnabledMods.ToArray() ) {
             if ( ! settings.TryGetValue( mod.Key, out ModSettings modSetting ) ) continue;
-            if ( ! modSetting.Disabled ) continue;
-            DisableAndRemoveMod( mod, "manual", "Mod {0} is manually disabled.", mod.Key );
+            if ( modSetting.Disabled )
+               DisableAndRemoveMod( mod, "manual", "Mod {0} is manually disabled.", mod.Key );
+            else if ( modSetting.LoadIndex.HasValue ) {
+               Log.Verbo( "Mod {0} LoadIndex manually set to {1}", mod.Key, modSetting.LoadIndex);
+               lock ( mod.Metadata ) mod.Metadata.LoadIndex = modSetting.LoadIndex.Value;
+            } else if ( modSetting.LogLevel.HasValue ) {
+               Log.Verbo( "Mod {0} LogLevel set to {1}", mod.Key, modSetting.LogLevel );
+               lock ( mod ) mod.LogLevel = modSetting.LogLevel.Value;
+            }
          }
       }